+++ cvat-ui/src/components/annotation-page/standard-workspace/controls-side-bar/control-visibility-observer.tsx
@@ -1,4 +1,5 @@
 // Copyright (C) 2021-2022 Intel Corporation
+// Copyright (C) 2022 CVAT.ai Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -8,6 +9,7 @@ import { SmallDashOutlined } from '@ant-design/icons';
 import Popover from 'antd/lib/popover';
 import React, { useEffect, useRef, useState } from 'react';
 import ReactDOM from 'react-dom';
+import { ConnectedComponent } from 'react-redux';
 
 const extraControlsContentClassName = 'cvat-extra-controls-control-content';
 
@@ -48,7 +50,7 @@ export function ExtraControlsControl(): JSX.Element {
 }
 
 export default function ControlVisibilityObserver<P = {}>(
-    ControlComponent: React.FunctionComponent<P>,
+    ControlComponent: React.FunctionComponent<P> | ConnectedComponent<any, any>,
 ): React.FunctionComponent<P> {
     let visibilityHeightThreshold = 0; // minimum value of height when element can be pushed to main panel