+++ cvat-ui/src/components/annotation-page/standard-workspace/objects-side-bar/object-item-menu.tsx
@@ -1,4 +1,5 @@
 // Copyright (C) 2020-2022 Intel Corporation
+// Copyright (C) 2022 CVAT.ai Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -6,7 +7,7 @@ import React from 'react';
 import Menu from 'antd/lib/menu';
 import Button from 'antd/lib/button';
 import Icon, {
-    LinkOutlined, CopyOutlined, BlockOutlined, RetweetOutlined, DeleteOutlined,
+    LinkOutlined, CopyOutlined, BlockOutlined, RetweetOutlined, DeleteOutlined, EditOutlined,
 } from '@ant-design/icons';
 
 import {
@@ -44,6 +45,7 @@ interface Props {
     toForeground(): void;
     resetCuboidPerspective(): void;
     changeColorPickerVisible(visible: boolean): void;
+    edit(): void;
     jobInstance: any;
 }
 
@@ -71,6 +73,20 @@ function MakeCopyItem(props: ItemProps): JSX.Element {
             <CVATTooltip title={`${copyShortcut} and ${pasteShortcut}`}>
                 <Button type='link' icon={<CopyOutlined />} onClick={copy}>
                     Make a copy
+                </Button>
+            </CVATTooltip>
+        </Menu.Item>
+    );
+}
+
+function EditMaskItem(props: ItemProps): JSX.Element {
+    const { toolProps, ...rest } = props;
+    const { edit } = toolProps;
+    return (
+        <Menu.Item {...rest}>
+            <CVATTooltip title='Shift + Double click'>
+                <Button type='link' icon={<EditOutlined />} onClick={edit}>
+                    Edit
                 </Button>
             </CVATTooltip>
         </Menu.Item>
@@ -209,6 +225,7 @@ export default function ItemMenu(props: Props): JSX.Element {
         TO_FOREGROUND = 'to_foreground',
         SWITCH_COLOR = 'switch_color',
         REMOVE_ITEM = 'remove_item',
+        EDIT_MASK = 'edit_mask',
     }
 
     const is2D = jobInstance.dimension === DimensionType.DIM_2D;
@@ -219,6 +236,7 @@ export default function ItemMenu(props: Props): JSX.Element {
             {!readonly && objectType !== ObjectType.TAG && (
                 <MakeCopyItem key={MenuKeys.COPY} toolProps={props} />
             )}
+            {!readonly && <EditMaskItem key={MenuKeys.EDIT_MASK} toolProps={props} />}
             {!readonly && <PropagateItem key={MenuKeys.PROPAGATE} toolProps={props} />}
             {is2D && !readonly && [ShapeType.POLYGON, ShapeType.POLYLINE, ShapeType.CUBOID].includes(shapeType) && (
                 <SwitchOrientationItem key={MenuKeys.SWITCH_ORIENTATION} toolProps={props} />