+++ cvat-ui/src/components/cvat-app.tsx
@@ -65,6 +65,8 @@ import showPlatformNotification, {
 import '../styles.scss';
 import EmailConfirmationPage from './email-confirmation-pages/email-confirmed';
 import EmailVerificationSentPage from './email-confirmation-pages/email-verification-sent';
+import IncorrectEmailConfirmationPage from './email-confirmation-pages/incorrect-email-confirmation';
+
 
 interface CVATAppProps {
     loadFormats: () => void;
@@ -429,6 +431,7 @@ class CVATApplication extends React.PureComponent<CVATAppProps & RouteComponentP
                     <Switch>
                         <Route exact path='/auth/register' component={RegisterPageContainer} />
                         <Route exact path='/auth/email-verification-sent' component={EmailVerificationSentPage} />
+                        <Route exact path='/auth/incorrect-email-confirmation' component={IncorrectEmailConfirmationPage} />
                         <Route exact path='/auth/login' component={LoginPageContainer} />
                         <Route
                             exact