+++ cvat/apps/dataset_manager/bindings.py
@@ -4,9 +4,11 @@
 #
 # SPDX-License-Identifier: MIT
 
+from functools import reduce
 import os.path as osp
 import re
 import sys
+import numpy as np
 from collections import namedtuple
 from pathlib import Path
 from types import SimpleNamespace
@@ -26,7 +28,7 @@ from cvat.apps.engine.models import Image as Img
 from cvat.apps.engine.models import Label, LabelType, Project, ShapeType, Task
 
 from .annotation import AnnotationIR, AnnotationManager, TrackManager
-from .formats.transformations import EllipsesToMasks
+from .formats.transformations import EllipsesToMasks, CVATRleToCOCORle
 
 CVAT_INTERNAL_ATTRIBUTES = {'occluded', 'outside', 'keyframe', 'track_id', 'rotation'}
 
@@ -1568,6 +1570,15 @@ def convert_cvat_anno_to_dm(cvat_frame_anno, label_attrs, map_label, format_name
                 "group": anno_group,
                 "attributes": anno_attr,
             }), cvat_frame_anno.height, cvat_frame_anno.width)
+        elif shape_obj.type == ShapeType.MASK:
+            anno = CVATRleToCOCORle.convert_mask(SimpleNamespace(**{
+                "points": shape_obj.points,
+                "label": anno_label,
+                "z_order": shape_obj.z_order,
+                "rotation": shape_obj.rotation,
+                "group": anno_group,
+                "attributes": anno_attr,
+            }), cvat_frame_anno.height, cvat_frame_anno.width)
         elif shape_obj.type == ShapeType.POLYLINE:
             anno = dm.PolyLine(anno_points,
                 label=anno_label, attributes=anno_attr, group=anno_group,
@@ -1671,7 +1682,8 @@ def import_dm_annotations(dm_dataset: dm.Dataset, instance_data: Union[ProjectDa
         dm.AnnotationType.polyline: ShapeType.POLYLINE,
         dm.AnnotationType.points: ShapeType.POINTS,
         dm.AnnotationType.cuboid_3d: ShapeType.CUBOID,
-        dm.AnnotationType.skeleton: ShapeType.SKELETON
+        dm.AnnotationType.skeleton: ShapeType.SKELETON,
+        dm.AnnotationType.mask: ShapeType.MASK
     }
 
     label_cat = dm_dataset.categories()[dm.AnnotationType.label]
@@ -1715,9 +1727,27 @@ def import_dm_annotations(dm_dataset: dm.Dataset, instance_data: Union[ProjectDa
                     points = []
                     if ann.type == dm.AnnotationType.cuboid_3d:
                         points = [*ann.position, *ann.rotation, *ann.scale, 0, 0, 0, 0, 0, 0, 0]
+                    elif ann.type == dm.AnnotationType.mask:
+                        istrue = np.argwhere(ann.image == 1).transpose()
+                        top = int(istrue[0].min())
+                        left = int(istrue[1].min())
+                        bottom = int(istrue[0].max())
+                        right = int(istrue[1].max())
+                        points = ann.image[top:bottom + 1, left:right + 1]
+
+                        def reduce_fn(acc, v):
+                            if v == acc['val']:
+                                acc['res'][-1] += 1
+                            else:
+                                acc['val'] = v
+                                acc['res'].append(1)
+                            return acc
+                        points = reduce(reduce_fn, points.reshape(np.prod(points.shape)), { 'res': [0], 'val': False })['res']
+                        points.extend([int(left), int(top), int(right), int(bottom)])
                     elif ann.type != dm.AnnotationType.skeleton:
                         points = ann.points
 
+                    rotation = ann.attributes.pop('rotation', 0.0)
                     # Use safe casting to bool instead of plain reading
                     # because in some formats return type can be different
                     # from bool / None
@@ -1765,6 +1795,7 @@ def import_dm_annotations(dm_dataset: dm.Dataset, instance_data: Union[ProjectDa
                             z_order=ann.z_order if ann.type != dm.AnnotationType.cuboid_3d else 0,
                             group=group_map.get(ann.group, 0),
                             source=source,
+                            rotation=rotation,
                             attributes=attributes,
                             elements=elements,
                         ))
@@ -1789,6 +1820,7 @@ def import_dm_annotations(dm_dataset: dm.Dataset, instance_data: Union[ProjectDa
                             points=points,
                             z_order=ann.z_order if ann.type != dm.AnnotationType.cuboid_3d else 0,
                             source=source,
+                            rotation=rotation,
                             attributes=attributes,
                         )